move dep_manager file
