+++ graphistry/tests/test_compute_cluster.py
@@ -5,7 +5,7 @@ import graphistry
 from graphistry.constants import DBSCAN
 from graphistry.util import ModelDict
 from graphistry.compute.cluster import lazy_dbscan_import_has_dependency
-from graphistry.dep_manager import deps
+from utils.dep_manager import deps
 
 has_dbscan, _, has_gpu_dbscan, _ = lazy_dbscan_import_has_dependency()
 has_umap = deps.umap
